final commit
set all the features

+++ RealTime_Ticketing_System/src/main/java/com/TicketingSystem/RealTime_Ticketing_System/Ticket.java
@@ -1,15 +1,18 @@
 package com.TicketingSystem.RealTime_Ticketing_System;
 
+import lombok.Generated;
+
 public class Ticket {
     private int ticketId;
     private String eventName;
-    private double ticketPrice;
+    private int ticketPrice;
 
-    public Ticket(int ticketId, String eventName, double ticketPrice){
+    public Ticket(int ticketId, String eventName, int ticketPrice){
         this.ticketId = ticketId;
         this.eventName = eventName;
         this.ticketPrice = ticketPrice;
     }
+
 
     public int getTicketId() {
         return ticketId;
@@ -31,7 +34,7 @@ public class Ticket {
         return ticketPrice;
     }
 
-    public void setTicketPrice(double ticketPrice) {
+    public void setTicketPrice(int ticketPrice) {
         this.ticketPrice = ticketPrice;
     }
